# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Property Management System
  version: 1.0.0
  description: |
    Source of truth for the boundary between the Angular application and the Go service.
    The Go server interface and the Angular client are both generated from this document;
    neither side hand-writes these types (Constitution III).

    Authorization is decided on the server for every operation. `x-required-role` states
    the requirement; an operation marked `admin` is refused for a manager regardless of
    how the request is made (Constitution I). Configuration — property types, areas,
    custom field definitions, and changing a property's reference code — are `admin`.
    Everything else is `manager`, which an administrator also satisfies.

    Sensitive custom fields never appear in this document as a filterable field. The
    server refuses a filter naming one even if a client constructs the request by hand
    (FR-027s4, Constitution VII).

    Every response body uses Western Arabic digits. Messages shown to users are Arabic.
    All timestamps are RFC 3339 in UTC; the client renders them in Africa/Cairo.
servers:
  - url: https://localhost:8443/api/v1
    description: Local development. TLS only — there is no plaintext listener.
...
security:
  - sessionCookie: []
paths:
...
  /properties:
    get:
      tags: [properties]
      operationId: listProperties
      summary: A page of the register
      description: |
        Sorted by name by default. `q` matches the name and the reference code after
        Arabic normalisation. `code` and `name` are per-column filters matching those
        single columns. Every filter combines by AND. Archived properties are excluded
        unless `includeArchived` is true.
      x-required-role: manager
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, enum: [10, 25, 50, 100], default: 25 }
        - name: q
          in: query
          description: Matched against name and reference code.
          schema: { type: string, maxLength: 120 }
        - name: code
          in: query
          description: |
            Column filter: substring match against the reference code after Arabic
            normalisation. Combines with every other filter by AND.
          schema: { type: string, maxLength: 32 }
        - name: name
          in: query
          description: |
            Column filter: substring match against the property name after Arabic
            normalisation. Combines with every other filter by AND.
          schema: { type: string, maxLength: 120 }
        - name: propertyTypeId
          in: query
          schema: { type: string, format: uuid }
        - name: areaId
          in: query
          schema: { type: string, format: uuid }
        - name: includeArchived
          in: query
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: A page of properties. An empty page is a valid answer.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PropertyPage' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }

    post:
      tags: [properties]
      operationId: createProperty
      summary: Add a property
      description: |
        `code` may be supplied only by an administrator; a manager supplying it is
        refused with 403. When omitted, the server generates the next code.
      x-required-role: manager
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PropertyCreate' }
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Property' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '409':
          description: The supplied reference code is already held, active or archived.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /properties/{propertyId}:
    parameters:
      - $ref: '#/components/parameters/PropertyId'
    get:
      tags: [properties]
      operationId: getProperty
      summary: One property with all of its fields
      x-required-role: manager
      responses:
        '200':
          description: The property, archived or not.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PropertyDetail' }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }

    put:
      tags: [properties]
      operationId: updateProperty
      summary: Modify a property
      description: |
        `version` must equal the version last read or the request is refused with
        `version_conflict`. The reference code is not part of this body; it has its own
        operation. An archived property is refused with `archived` until restored.
      x-required-role: manager
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PropertyUpdate' }
      responses:
        '200':
          description: Updated, with the new version.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PropertyDetail' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409':
          description: |
            The property changed since it was read, or it is archived.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: pms_session
      description: |
        Set by /auth/login. HttpOnly, Secure, SameSite=Strict. Not readable by
        JavaScript and never placed in a URL. Carries identity only; role and active
        state are read from stored data on every request.
  parameters:
...
    PropertyId:
      name: propertyId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
...
    PropertyCode:
      type: string
      minLength: 1
      maxLength: 32
      description: |
        Unique across the register, active and archived alike, compared after
        trimming, case folding, and Arabic normalisation. Never reused.
    PropertyName:
      type: string
      minLength: 2
      maxLength: 120
      description: |
        Descriptive, not identifying. Two properties may share a name.
    Lookup:
      type: object
      required: [id, label]
      properties:
        id:    { type: string, format: uuid }
        label: { type: string, minLength: 1, maxLength: 60 }
...
    CustomFieldValue:
      type: object
      required: [fieldId]
      description: |
        Exactly one of the value members is present, matching the field's type. A
        sensitive field's value arrives here decrypted, and only on the property
        detail — never in a list or a search result.
      properties:
        fieldId:   { type: string, format: uuid }
        text:      { type: [string, 'null'] }
        checked:   { type: [boolean, 'null'] }
        choiceId:  { type: [string, 'null'], format: uuid }
        choiceIds:
          type: array
          items: { type: string, format: uuid }
    PropertySummary:
      type: object
      required: [id, code, name, propertyType, area, isArchived]
      description: One row of the list. Custom values are not included here.
      properties:
        id:           { type: string, format: uuid }
        code:         { $ref: '#/components/schemas/PropertyCode' }
        name:         { $ref: '#/components/schemas/PropertyName' }
        propertyType: { $ref: '#/components/schemas/Lookup' }
        area:         { $ref: '#/components/schemas/Lookup' }
        isArchived:   { type: boolean }
    Property:
      allOf:
        - $ref: '#/components/schemas/PropertySummary'
        - type: object
          required: [version, createdAt, createdBy, updatedAt, updatedBy]
          properties:
            version:    { type: integer }
            createdAt:  { type: string, format: date-time }
            createdBy:  { type: string }
            updatedAt:  { type: string, format: date-time }
            updatedBy:  { type: string }
            archivedAt: { type: [string, 'null'], format: date-time }
            archivedBy: { type: [string, 'null'] }
    PropertyDetail:
      allOf:
        - $ref: '#/components/schemas/Property'
        - type: object
          required: [customValues]
          properties:
            customValues:
              type: array
              items: { $ref: '#/components/schemas/CustomFieldValue' }
    PropertyCreate:
      type: object
      required: [name, propertyTypeId, areaId]
      properties:
        name:           { $ref: '#/components/schemas/PropertyName' }
        propertyTypeId:
          type: string
          format: uuid
          description: Must name a configured type; a free-typed value is impossible.
        areaId:         { type: string, format: uuid }
        code:
          $ref: '#/components/schemas/PropertyCode'
          description: |
            Administrator only. A manager supplying it is refused with 403. Omitted,
            the server generates the next code.
        customValues:
          type: array
          items: { $ref: '#/components/schemas/CustomFieldValue' }
    PropertyUpdate:
      type: object
      required: [name, propertyTypeId, areaId, version]
      properties:
        name:           { $ref: '#/components/schemas/PropertyName' }
        propertyTypeId: { type: string, format: uuid }
        areaId:         { type: string, format: uuid }
        version:        { type: integer }
        customValues:
          type: array
          items: { $ref: '#/components/schemas/CustomFieldValue' }
    PropertyPage:
      type: object
      required: [items, page, pageSize, totalItems]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/PropertySummary' }
        page:       { type: integer }
        pageSize:   { type: integer }
        totalItems: { type: integer }
...
    Error:
      type: object
      required: [code, message]
      description: |
        The single error shape used by every operation in this document
        (Constitution III). `message` is Arabic and states what the user should do
        next. It never carries internal failure detail, a key, a wrapped key, or a
        decrypted value (Constitution VII).
      properties:
        code:
          type: string
          enum:
            - invalid_request
            - invalid_credentials
            - too_soon
            - not_authenticated
            - forbidden
            - not_found
            - conflict
            - in_use
            - version_conflict
            - archived
            - password_change_required
            - internal_error
        message:
          type: string
          description: Arabic, user-facing.
        fields:
          type: [object, 'null']
          additionalProperties:
            type: string
          description: Field-level Arabic messages, keyed by field name, for form errors.
        retryAfterSeconds:
          type: [integer, 'null']
          minimum: 1
          maximum: 60
          description: Seconds remaining before another attempt is accepted (sign-in throttling).
  responses:
    Error:
      description: The request was refused. Body is the single error shape.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'